// Help247/HelpWeb/src/app/auth/models/register.model.ts
export class RegisterModel {
    userName: string
    password: string
    userType: number
    email: string
}


export class CustomerRegisterModel {
    userName: string
    password: string
    userType: number
    phoneNo: string
    email: string
    country: string
    state: string
    city: string
    postalCode: string
    addressLine: string
    profilePicUrl: string
    name: string
}

export class HelperRegisterModel {
    userName: string;
    password: string;
    userType: number;
    firstName: string;
    lastName: string;
    phoneNo: string;
    mobileNo: string;
    email: string;
    addressLine: string;
    country: string;
    state: string;
    city: string;
    postalCode: string;
    experience: number;
    profilePicUrl: string;
    aboutMe: string;
    myService: string;
    helperCategoryId: 0;
}